Added retry logic to transactions when database is locked (#38)
Test suite in `dwn-sdk-js` introduced a test that has concurrent
operations in v0.3.7.

That test fails in SQLite implementations due to the database being
locked. Introduced retry logic + reduced the default timeout of 5
seconds before attempting the query (and fail if still locked) when
database is locked to address this issue.